refactor(calendar): simplify month navigation buttons using `useCalendarSetMonthButton`
