+++ packages/react/src/calendar/decrement-month/CalendarDecrementMonth.tsx
@@ -1,14 +1,7 @@
 'use client';
 import * as React from 'react';
-import { useStore } from '@base-ui/utils/store';
-import { useSharedCalendarRootContext } from '../root/SharedCalendarRootContext';
-import { useRenderElement } from '../../utils/useRenderElement';
 import { BaseUIComponentProps, NativeButtonProps } from '../../utils/types';
-import { useButton } from '../../use-button';
-import { useTemporalAdapter } from '../../temporal-adapter-provider/TemporalAdapterContext';
-import { selectors } from '../store';
-import { useCalendarMonthButton } from '../utils/useCalendarMonthButton';
-import { REASONS } from '../../utils/reasons';
+import { useCalendarSetMonthButton } from '../utils/useCalendarSetMonthButton';
 
 /**
  * Displays an element to navigate to the previous month in the calendar.
@@ -21,72 +14,12 @@ export const CalendarDecrementMonth = React.forwardRef(function CalendarDecremen
   componentProps: CalendarDecrementMonth.Props,
   forwardedRef: React.ForwardedRef<HTMLButtonElement>,
 ) {
-  const {
-    className,
-    render,
-    nativeButton,
-    disabled: disabledProp,
-    ...elementProps
-  } = componentProps;
-
-  const store = useSharedCalendarRootContext();
-  const adapter = useTemporalAdapter();
-  const monthPageSize = useStore(store, selectors.monthPageSize);
-  const visibleDate = useStore(store, selectors.visibleDate);
-
-  const targetDate = React.useMemo(
-    () => adapter.addMonths(visibleDate, -monthPageSize),
-    [visibleDate, monthPageSize, adapter],
-  );
-
-  const isDisabled = useStore(store, selectors.isSetMonthButtonDisabled, targetDate, disabledProp);
-
-  const { getButtonProps, buttonRef } = useButton({
-    disabled: isDisabled,
-    native: nativeButton,
-    focusableWhenDisabled: true,
-  });
-
-  const { pointerHandlers, autoChangeButtonRef, shouldSkipClick } = useCalendarMonthButton({
+  return useCalendarSetMonthButton({
     direction: -1,
-    disabled: isDisabled,
-    disabledProp,
-    store,
-    adapter,
-    monthPageSize,
+    ariaLabel: { singular: 'Previous month', plural: 'Previous months' },
+    componentProps,
+    forwardedRef,
   });
-
-  const state: CalendarDecrementMonth.State = React.useMemo(
-    () => ({ disabled: isDisabled }),
-    [isDisabled],
-  );
-
-  const element = useRenderElement('button', componentProps, {
-    state,
-    ref: [buttonRef, autoChangeButtonRef, forwardedRef],
-    props: [
-      {
-        tabIndex: 0,
-        'aria-label': monthPageSize > 1 ? 'Previous months' : 'Previous month',
-        onClick(event) {
-          if (isDisabled || shouldSkipClick(event)) {
-            return;
-          }
-          store.setVisibleDate(
-            targetDate,
-            event.nativeEvent,
-            event.currentTarget as HTMLElement,
-            REASONS.monthChange,
-          );
-        },
-        ...pointerHandlers,
-      },
-      elementProps,
-      getButtonProps,
-    ],
-  });
-
-  return element;
 });
 
 export interface CalendarDecrementMonthState {
