+++ packages/react/src/calendar/increment-month/CalendarIncrementMonth.tsx
@@ -1,14 +1,7 @@
 'use client';
 import * as React from 'react';
-import { useStore } from '@base-ui/utils/store';
-import { useSharedCalendarRootContext } from '../root/SharedCalendarRootContext';
-import { useRenderElement } from '../../utils/useRenderElement';
 import { BaseUIComponentProps, NativeButtonProps } from '../../utils/types';
-import { useButton } from '../../use-button';
-import { useTemporalAdapter } from '../../temporal-adapter-provider/TemporalAdapterContext';
-import { selectors } from '../store';
-import { useCalendarMonthButton } from '../utils/useCalendarMonthButton';
-import { REASONS } from '../../utils/reasons';
+import { useCalendarSetMonthButton } from '../utils/useCalendarSetMonthButton';
 
 /**
  * Displays an element to navigate to the next month in the calendar.
@@ -20,72 +13,12 @@ export const CalendarIncrementMonth = React.forwardRef(function CalendarIncremen
   componentProps: CalendarIncrementMonth.Props,
   forwardedRef: React.ForwardedRef<HTMLButtonElement>,
 ) {
-  const {
-    className,
-    render,
-    nativeButton,
-    disabled: disabledProp,
-    ...elementProps
-  } = componentProps;
-
-  const store = useSharedCalendarRootContext();
-  const adapter = useTemporalAdapter();
-  const monthPageSize = useStore(store, selectors.monthPageSize);
-  const visibleDate = useStore(store, selectors.visibleDate);
-
-  const targetDate = React.useMemo(
-    () => adapter.addMonths(visibleDate, monthPageSize),
-    [visibleDate, monthPageSize, adapter],
-  );
-
-  const isDisabled = useStore(store, selectors.isSetMonthButtonDisabled, targetDate, disabledProp);
-
-  const { getButtonProps, buttonRef } = useButton({
-    disabled: isDisabled,
-    native: nativeButton,
-    focusableWhenDisabled: true,
-  });
-
-  const { pointerHandlers, autoChangeButtonRef, shouldSkipClick } = useCalendarMonthButton({
+  return useCalendarSetMonthButton({
     direction: 1,
-    disabled: isDisabled,
-    disabledProp,
-    store,
-    adapter,
-    monthPageSize,
+    ariaLabel: { singular: 'Next month', plural: 'Next months' },
+    componentProps,
+    forwardedRef,
   });
-
-  const state: CalendarIncrementMonth.State = React.useMemo(
-    () => ({ disabled: isDisabled }),
-    [isDisabled],
-  );
-
-  const element = useRenderElement('button', componentProps, {
-    state,
-    ref: [buttonRef, autoChangeButtonRef, forwardedRef],
-    props: [
-      {
-        tabIndex: 0,
-        'aria-label': monthPageSize > 1 ? 'Next months' : 'Next month',
-        onClick(event) {
-          if (isDisabled || shouldSkipClick(event)) {
-            return;
-          }
-          store.setVisibleDate(
-            targetDate,
-            event.nativeEvent,
-            event.currentTarget as HTMLElement,
-            REASONS.monthChange,
-          );
-        },
-        ...pointerHandlers,
-      },
-      elementProps,
-      getButtonProps,
-    ],
-  });
-
-  return element;
 });
 
 export interface CalendarIncrementMonthState {
